A Pure Photo Gallery
In this project I created a photo gallery utilizing only HTML5 and CSS3.
This is based on a project I did while at The Tech Academy. I changed the UI during the project to learn theming and cohesion.
I also tried a glow animation for the first time on the navigation buttons.