Single Page Application using KnockoutJS
This project was used to learn KnockoutJS. I created a non functioning pizza menu, with a customized theme and layout in order to create a pizza. This was done as an ASP.NET Single Page Application. Though you do not get a price or even delivery, you get to pretend to build a pizza!